Упростил плеер и вынес состояние карточки в HOC
